Experimental fixes for Bucket data race conditions
If multiple go-routines are using a bucket and a NOT_MY_VBUCKET
error is detected, the Bucket.refresh operation is started to update
the clients mapping of vbucket ids to servers.

However, the process is not thread safe and causes intermittent probelems
with partial updates in a heavy load environemnt (many goroutines operating
on a bucket that is rebalancing itself of the shard).

Not happy about the use of RWMutex, but it's a stop-gap solution for the time
being.

This also includes a change to handle the case where a couchbase pool
advertises a bucket with an empty server list. I believe this only occurs
during server warmup and is a temporary issue. Instead of crashing on an
index-out-of-bounds error, we simply refresh the bucket state. This maybe
needs a timeout mechanism for general use, but the infinite retry is good
enough for my needs at the time.