Update the sleep delay with server estimate
Another change I'm a bit unsure of. WaitForPersistence almost universally
fails if the server returns an average PersistenceTime. Instead of adjusting
the total timeout to 2x the estimate, we adjust the sleepDelay to the
estimate. I haven't experimented enough to be happy with this solution; maybe
we should try something like half or three-quartes the PersistenceTime (not
sure how often in production the persistence time varies significantly from
the average). I also feel the total timeout should be adjusted, but I'm not
really confident on how that adjustment should look yet.