Fix TAP connections running in KeysOnly+Dump mode
These connections are closed by the server when the list of keys has
been sent to the client. This change updates the automatic connection
retry to forward the memcache feed error to the TapFeed.run process.
If the memcache client did not report an erorr with the stream, we close
the feed and exit gracefully. Otherwise, we trigger the connection retry
logic as before.